기능: IAP sandbox 환경 구성 + Sentry → Crashlytics 전환
- iOS entitlements에 In-App Purchase capability 추가
- StoreKit Configuration 파일 생성 (시뮬레이터 인프라 검증용)
- Xcode Scheme에 StoreKit Configuration 연결
- envied 패키지로 .env 기반 API 키 관리 전환
  - RevenueCat iOS/Android 키 플랫폼별 자동 분기
  - dart_defines.json 방식 제거
- SubscriptionService: API 키 빈 문자열 시 크래시 방어
- Sentry 완전 제거 → Firebase Crashlytics 전환
  - ErrorHandler, AnalyticsService, main.dart, app.dart 전환
  - sentry_scrubber.dart 및 테스트 삭제
  - CI 워크플로우에서 Sentry 시크릿 제거

+++ test/data/services/reminder_service_test.mocks.dart
@@ -390,15 +390,15 @@ class MockTimezoneService extends _i1.Mock implements _i12.TimezoneService {
           as _i2.Location);
 
   @override
-  double getTimeDifference(String? homeTimezone, String? currentTimezone) =>
+  int getTimeDifference(String? homeTimezone, String? currentTimezone) =>
       (super.noSuchMethod(
             Invocation.method(#getTimeDifference, [
               homeTimezone,
               currentTimezone,
             ]),
-            returnValue: 0.0,
+            returnValue: 0,
           )
-          as double);
+          as int);
 
   @override
   DateTime convertTime(
